consolidate definitions of print_header_line
