+++ helpers.py
@@ -4,3 +4,17 @@ def sizeof_fmt(num):
             return "%3.1f%s" % (num, x)
         num /= 1024.0
 
+def print_header_line():
+    l = [ "Title",
+                "Package name",
+                "Creator",
+                "Super Dev",
+                "Price",
+                "Offer Type",
+                "Version Code",
+                "Size",
+                "Rating",
+                "Num Downloads",
+             ]
+    print SEPARATOR.join(l)
+
